+++ drivers/net/ethernet/intel/ice/ice_type.h
@@ -258,7 +258,7 @@ enum ice_fd_hw_seg {
 struct ice_fd_hw_prof {
 	struct ice_flow_seg_info *fdir_seg[ICE_FD_HW_SEG_MAX];
 	int cnt;
-	u64 entry_h[ICE_MAX_FDIR_VSI_PER_FILTER][ICE_FD_HW_SEG_MAX];
+	uintptr_t entry_h[ICE_MAX_FDIR_VSI_PER_FILTER][ICE_FD_HW_SEG_MAX];
 	u16 vsi_h[ICE_MAX_FDIR_VSI_PER_FILTER];
 	u64 prof_id[ICE_FD_HW_SEG_MAX];
 };
